requirements.txt
proxmoxer==2.0.1
PyYAML==6.0
requests==2.28.2